+++ R/RcppExports.R
@@ -9,6 +9,12 @@ computeLinear <- function(curValue, target, x, w, boundLinear = 10) {
 
 #' @rdname computeFrac
 #' @export
+computeLinearG1_old <- function(curValue, target, x, w, boundLinear = 10) {
+    .Call('_surveysd_computeLinearG1_old', PACKAGE = 'surveysd', curValue, target, x, w, boundLinear)
+}
+
+#' @rdname computeFrac
+#' @export
 computeLinearG1 <- function(curValue, target, x, w, boundLinear = 10) {
     .Call('_surveysd_computeLinearG1', PACKAGE = 'surveysd', curValue, target, x, w, boundLinear)
 }