+++ R/plot.R
@@ -30,10 +30,9 @@
 #' @examples
 #'
 #' library(surveysd)
-#' library(laeken)
-#' library(data.table)
-#'
-#' eusilc <- demo.eusilc(n = 4, prettyNames = TRUE)
+#' 
+#' set.seed(1234)
+#' eusilc <- demo.eusilc(n = 3, prettyNames = TRUE)
 #'
 #' dat_boot <- draw.bootstrap(eusilc, REP = 3, hid = "hid", weights = "pWeight",
 #'                            strata = "region", period = "year")
@@ -54,15 +53,11 @@
 #' # dotted line is the result on the national level
 #' plot(err.est, type = "grouping", groups = "gender")
 #'
-#' # plot results for gender
-#' # with standard errors as ribbons
-#' plot(err.est, type = "grouping", groups = "gender", sd.type = "ribbon")
 #'
 #' # plot results for rb090 in each db040
-#' plot(err.est, type = "grouping", groups = c("gender", "region"))
+#' # with standard errors as ribbons
+#' plot(err.est, type = "grouping", groups = c("gender", "region"), sd.type = "ribbon")
 #'
-#' # plot results for db040 in each rb090 with standard errors as ribbons
-#' plot(err.est,type = "grouping", groups = c("gender", "region"))
 #'
 #' @export
 
@@ -354,8 +349,8 @@ plot.surveysd <- function(
     if (groups[1] == period) {
       p1 <- p1 + facet_wrap(~get(groups[2]))
       if (sd.type == "ribbon") {
-        p1 <- p1 + guides(fill = FALSE,
-                          colour = FALSE,
+        p1 <- p1 + guides(fill = 'none',
+                          colour = 'none',
                           linetype = guide_legend(title = ""))
         p1 <- p1 + scale_linetype_manual(
           values = c("solid"), labels = paste0("Result for ", groups[1]))
@@ -363,7 +358,7 @@ plot.surveysd <- function(
         shapes <- c(1, 16)
         names(shapes) <- c("high st.Error", "high st.Error for mean")
         p1 <- p1 + scale_shape_manual(values = shapes)
-        p1 <- p1 + scale_colour_discrete(guide = FALSE) +
+        p1 <- p1 + scale_colour_discrete(guide = 'none') +
         # p1 <- p1 + scale_colour_discrete(value="grey")
            theme(legend.title = element_blank())
         p1 <- p1 + scale_linetype_manual(values = c("dotted"), labels = paste0(